requirements.txt
beautifulsoup4==4.12.3
cfgv==3.4.0
django-cors-headers==4.3.1
django-crequest==2018.5.11
django-dotenv==1.4.2
django-extensions==3.2.3
django-filter==23.5
django-ninja==1.3.0
django-oauth-toolkit==2.3.0
drf-spectacular==0.27.2
identify==2.5.33
nodeenv==1.8.0
packaging==23.2
pillow==10.3.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
tzdata==2023.4
virtualenv==20.25.0
